fix:my-scripts
alocating all scripts to one place
