--- 03-back-end/bloco-22-introducao-ao-desenvolvimento-web-com-nodejs/dia-01-intro-nodejs/22.1-exercises/avg-car-speed/index.js
@@ -1,8 +1,0 @@
-const speed = require('./speed');
-const { questionInt } = require('readline-sync');
-
-console.log('Measuring the car average speed')
-const distance = questionInt('What is the traveled distance in meters? ');
-const time = questionInt('It took how much time to complete the path in minutes? ');
-
-speed(distance, time);
