--- 03-back-end/bloco-22-introducao-ao-desenvolvimento-web-com-nodejs/dia-01-intro-nodejs/22.1-exercises/avg-car-speed/speed.js
@@ -1,6 +1,0 @@
-const avgSpeed = (distance, minutesTime) => {
-  const avgResult = distance / (minutesTime * 60);
-  console.log(`With the distance of ${distance} meters and the ${minutesTime} minutes time the average of speed was ${avgResult.toFixed(2)} seconds per meters`);
-};
-
-module.exports = avgSpeed;
