--- 03-back-end/bloco-22-introducao-ao-desenvolvimento-web-com-nodejs/dia-01-intro-nodejs/22.1-exercises/calculator/index.js
@@ -1,9 +1,0 @@
-const { question, questionInt } = require('readline-sync');
-const calculator = require('./calculator');
-
-console.log('Basic calculator');
-const num1 = questionInt('Insert the first number to calculate ');
-const operator = question('Insert the operator ');
-const num2 = questionInt('Insert the second number ');
-
-calculator(num1, num2, operator);
