--- 03-back-end/bloco-22-introducao-ao-desenvolvimento-web-com-nodejs/dia-01-intro-nodejs/22.1-exercises/calculator/package.json
@@ -1,16 +1,0 @@
-{
-  "name": "calculator",
-  "version": "1.0.0",
-  "description": "basic calculation by input",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node index.js"
-  },
-  "keywords": [],
-  "author": "Davíd Roggér",
-  "license": "ISC",
-  "dependencies": {
-    "readline-sync": "^1.4.10"
-  }
-}
