--- 03-back-end/bloco-22-introducao-ao-desenvolvimento-web-com-nodejs/dia-01-intro-nodejs/22.1-exercises/guessing-game/index.js
@@ -1,5 +1,0 @@
-const game = require('./prize-draw');
-
-console.table('Welcome to the guessing game!');
-
-game();
