+++ 03-back-end/bloco-22-introducao-ao-desenvolvimento-web-com-nodejs/dia-01-intro-nodejs/22.1-exercises/my-scripts/calculator.js
@@ -1,3 +1,5 @@
+const { question, questionInt } = require('readline-sync');
+
 const calculator = (num1, num2, operator) => {
   switch(operator) {
     case "+":
@@ -16,5 +18,12 @@ const calculator = (num1, num2, operator) => {
       console.log('Insert a valid operator');
   }
 }
+
+console.log('Basic calculator');
+const num1 = questionInt('Insert the first number to calculate ');
+const operator = question('Insert the operator ');
+const num2 = questionInt('Insert the second number ');
+
+calculator(num1, num2, operator);
 
 module.exports = calculator;
