+++ 03-back-end/bloco-22-introducao-ao-desenvolvimento-web-com-nodejs/dia-01-intro-nodejs/22.1-exercises/my-scripts/imc.js
@@ -1,3 +1,5 @@
+const { questionFloat } = require('readline-sync');
+
 const imcCalculation = (weight, height) => {
   const imcResult = weight / (height ^ 2);
   console.log(`With the height of ${height} meters and the weight of ${weight} kilograms your IMC is ${imcResult}`);
@@ -7,7 +9,14 @@ const imcCalculation = (weight, height) => {
   if(imcResult >= 30.0 && imcResult <= 34.9) console.log('Obesity grade one.');
   if(imcResult >= 35.0 && imcResult <= 39.9) console.log('Obesity grade two.');
   if(imcResult > 40.0) console.log('Obesity grade two and four.');
-
+  
 };
 
-module.exports = imcCalculation;
+const imcQuestions = () => {
+  const weight = questionFloat('Please insert your weight like in this quote "70.0"KG ');
+  const height = questionFloat('Now insert your height like in this quote "1.70"meters ');
+  
+  imcCalculation(weight, height);
+};
+
+module.exports = imcQuestions;
