+++ 03-back-end/bloco-22-introducao-ao-desenvolvimento-web-com-nodejs/dia-01-intro-nodejs/22.1-exercises/my-scripts/package.json
@@ -5,10 +5,14 @@
   "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "imc": "node index.js"
+    "start": "node index.js",
+    "imc": "node imc.js",
+    "calculator": "node calculator.js",
+    "speed": "node speed.js",
+    "raffle": "node prize-draw.js"
   },
   "keywords": [],
-  "author": "",
+  "author": "Davíd Roggér",
   "license": "ISC",
   "dependencies": {
     "readline-sync": "^1.4.10"
