+++ 03-back-end/bloco-22-introducao-ao-desenvolvimento-web-com-nodejs/dia-01-intro-nodejs/22.1-exercises/my-scripts/prize-draw.js
@@ -20,6 +20,7 @@ const playerAnswerCheck = (answer) => {
 }
 
 const game = () => {
+  console.table('Welcome to the guessing game!');
   while (playGame.start) {
     playGame.keepPlayingQuestion = true;
     guessingGame()
